Expose `apply` to allow arbitrary unitary application
This updates the sparse simulator to make public the `apply` function used in tests that allows for (slow but stable) application of arbitrary unitaries to the current sparse state.
